% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rank.R
\name{clean_ranks}
\alias{clean_ranks}
\title{Clean messy ranks into simple ranks, converting
ranks with more than one breeding status (e.g., "S5N,S2B") into
single ranks (specified by the \code{keep} paramater)}
\usage{
clean_ranks(ranks, keep = "B")
}
\arguments{
\item{ranks}{a character vector of ranks}

\item{keep}{which component to of a rank with multiple ranks for different
breeding statuses. \code{"B"} = breeding, \code{"M"} = migratory, \code{"N"} = non-breeding}
}
\value{
character vector of cleaned ranks
}
\description{
Clean messy ranks into simple ranks, converting
ranks with more than one breeding status (e.g., "S5N,S2B") into
single ranks (specified by the \code{keep} paramater)
}
