% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rank.R
\name{ranks_to_numeric}
\alias{ranks_to_numeric}
\title{Convert S ranks to numeric}
\usage{
ranks_to_numeric(ranks, simplify = TRUE)
}
\arguments{
\item{ranks}{character vector of input S ranks}
}
\value{
a list the same length as \code{ranks} of numeric vectors. For range ranks,
the vector will be sequence from low to high (e.g.,\code{"S3S5"} becomes \code{c(3,4,5)})
}
\description{
Converts an rank with S to a single number
}
\examples{
ranks_to_numeric(c("S1","S4","S3S4"))

}
