% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_rank_group.R
\name{sampled_index}
\alias{sampled_index}
\title{Function to automate the rank trends by taxanomic group}
\usage{
sampled_index(wt_data, tax_group, wts_col, yr_col, nreps = 1000)
}
\arguments{
\item{wt_data}{input dataframe of species weighted ranks}

\item{tax_group}{name of column that contains grouping variable e.g. "taxonomic group"}

\item{wts_col}{name of column with numeric weight values (e.g. "wts_col")}

\item{yr_col}{name of column that specifies assessment (e.g. "year")}

\item{nreps}{number of times to subsample species weights (default = 1000)}
}
\value{
a dataframe containing mean, upper and lower confidence interval for samples
conservation status index
}
\description{
This function generates confidence intervals for ranks by taxanomic groups by resampling
species within a given group at each time point.
}
